Initial commit, testing PR
This branch was based on master before I made a change to bar.txt
directly on master using the GitHub web interface.
After making the branch, I fixed some typos in foo.txt and added a line.
I also copied the current contents of bar.txt from the GitHub web
interface and pasted them into bar.txt for this branch.
I want to see if the changes to bar.txt show up in the PR.  I think they
should not, since the contents of bar.txt in this branch are now
identical to those in the current version of master.